moved information json files to assets
