+++ src/assets/information-data/termsInformation.json
