+++ src/screens/Profile/Information/TermsInformation.tsx
@@ -1,6 +1,7 @@
 import {InformationTexts, useTranslation} from '@atb/translations';
 import React from 'react';
-import termsInfo from '../../../reference-data/termsInformation.json';
+import termsInfo from '../../../assets/information-data/termsInformation.json';
+
 import Information, {InformationElement} from './Information';
 
 export default function TermsInformation() {
